move database models to  file
