+++ app.py
@@ -1,11 +1,11 @@
-from flask import Flask, render_template, url_for, session, redirect
-from flask_sqlalchemy import SQLAlchemy
+from flask import Flask, render_template
+from models import db
 import blueprints.auth as auth
 
 # start and configure app
 app = Flask(__name__)
 app.config.from_pyfile('app.cfg')
-
+db.init_app(app)
 
 # register blueprints
 app.register_blueprint(auth.auth)
@@ -16,22 +16,6 @@ auth.configure_oauth(
     app.config['GOOGLE_OAUTH_CLIENTID'],
     app.config['GOOGLE_OAUTH_SECRET'],
     app.config['SERVER_NAME'])
-db = SQLAlchemy(app)
-
-# Database Models
-
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True)
-    email = db.Column(db.String(120), unique=True)
-
-    def __init__(self, username, email):
-        self.username = username
-        self.email = email
-
-    def __repr__(self):
-        return '<User %r>' % self.username
 
 # Routes
