+++ init-db.py
@@ -1,3 +1,4 @@
 if __name__ == '__main__':
-    from app import db
-    db.create_all()
+    from app import app, db
+    with app.app_context():
+        db.create_all()
